requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
Django==5.1.3
django-modeltranslation==0.19.11
environs==11.2.1
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
marshmallow==3.23.1
multidict==6.1.0
numpy==2.1.3
openpyxl==3.1.5
orjson==3.10.12
packaging==24.2
pandas==2.2.3
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
yarl==1.18.0
